producer/requirements.txt
kafka-python
feedparser
aiohttp